+++ apps/client/src/features/control/playback/TimerControlExport.tsx
@@ -1,7 +1,7 @@
 import { Box } from '@chakra-ui/react';
 import { IoArrowUp } from '@react-icons/all-files/io5/IoArrowUp';
 
-import ErrorBoundary from '../../../common/components/errorBoundary/ErrorBoundary';
+import ErrorBoundary from '../../../common/components/error-boundary/ErrorBoundary';
 import { handleLinks } from '../../../common/utils/linkUtils';
 
 import PlaybackControl from './PlaybackControl';
