+++ apps/client/src/features/editors/ProtectedEditor.jsx
@@ -1,4 +1,4 @@
-import ProtectRoute from '../../common/components/protectRoute/ProtectRoute';
+import ProtectRoute from '../../common/components/protect-route/ProtectRoute';
 
 import Editor from './Editor';
